crux_http::middleware
Struct Redirect

pub struct Redirect { /* private fields */ }

A middleware which attempts to follow HTTP redirects.

Implementations

impl Redirect

pub fn new(attempts: u8) -> Self

Create a new instance of the Redirect middleware, which attempts to follow redirects up to as many times as specified.

Consider using Redirect::default() for the default number of redirect attempts.

This middleware will follow redirects from the Location header if the server returns any of the following http response codes:

301 Moved Permanently

302 Found

303 See other

307 Temporary Redirect

308 Permanent Redirect

Errors

An error will be passed through the middleware stack if the value of the Location header is not a validly parsing url.

Caveats

This will presently make at least one additional HTTP request before the actual request to determine if there is a redirect that should be followed, so as to preserve any request body.

Examples

caps.http
    .get("https://httpbin.org/redirect/2")
    .middleware(crux_http::middleware::Redirect::default())
    .send(Event::ReceiveResponse)