Struct crux_http::Request

pub struct Request { /* private fields */ }

An HTTP request, returns a Response.

Implementations

impl Request

pub fn new(method: Method, url: Url) -> Self

Create a new instance.

This method is particularly useful when input URLs might be passed by third parties, and you don’t want to panic if they’re malformed. If URLs are statically encoded, it might be easier to use one of the shorthand methods instead.

Examples

fn main() -> crux_http::Result<()> {
use crux_http::http::{Url, Method};

let url = Url::parse("https://httpbin.org/get")?;
let req = crux_http::Request::new(Method::Get, url);

pub fn query<T: DeserializeOwned>(&self) -> Result<T>

Get the URL querystring.

Examples

#[derive(Serialize, Deserialize)]
struct Index {
    page: u32
}

let req = caps.http.get("https://httpbin.org/get?page=2").build();
let Index { page } = req.query()?;
assert_eq!(page, 2);

pub fn set_query(&mut self, query: &impl Serialize) -> Result<()>

Set the URL querystring.

Examples

#[derive(Serialize, Deserialize)]
struct Index {
    page: u32
}

let query = Index { page: 2 };
let mut req = caps.http.get("https://httpbin.org/get").build();
req.set_query(&query)?;
assert_eq!(req.url().query(), Some("page=2"));
assert_eq!(req.url().as_str(), "https://httpbin.org/get?page=2");

pub fn header(&self, key: impl Into<HeaderName>) -> Option<&HeaderValues>

Get an HTTP header.

Examples

let mut req = caps.http.get("https://httpbin.org/get").build();
req.set_header("X-Requested-With", "surf");
assert_eq!(req.header("X-Requested-With").unwrap(), "surf");

pub fn header_mut(
&mut self,
name: impl Into<HeaderName>
) -> Option<&mut HeaderValues>

Get a mutable reference to a header.

pub fn insert_header(
&mut self,
name: impl Into<HeaderName>,
values: impl ToHeaderValues
) -> Option<HeaderValues>

Set an HTTP header.

pub fn append_header(
&mut self,
name: impl Into<HeaderName>,
values: impl ToHeaderValues
)

Append a header to the headers.

Unlike insert this function will not override the contents of a header, but insert a header if there aren’t any. Or else append to the existing list of headers.

pub fn remove_header(
&mut self,
name: impl Into<HeaderName>
) -> Option<HeaderValues>

Remove a header.

pub fn iter(&self) -> Iter<'_>

An iterator visiting all header pairs in arbitrary order.

pub fn iter_mut(&mut self) -> IterMut<'_>

An iterator visiting all header pairs in arbitrary order, with mutable references to the values.

pub fn header_names(&self) -> Names<'_>

An iterator visiting all header names in arbitrary order.

pub fn header_values(&self) -> Values<'_>

An iterator visiting all header values in arbitrary order.

pub fn set_header(
&mut self,
key: impl Into<HeaderName>,
value: impl ToHeaderValues
)

Set an HTTP header.

Examples

let mut req = caps.http.get("https://httpbin.org/get").build();
req.set_header("X-Requested-With", "surf");
assert_eq!(req.header("X-Requested-With").unwrap(), "surf");

pub fn ext<T: Send + Sync + 'static>(&self) -> Option<&T>

Get a request extension value.

pub fn set_ext<T: Send + Sync + 'static>(&mut self, val: T) -> Option<T>

Set a request extension value.

pub fn method(&self) -> Method

Get the request HTTP method.

Examples

let req = caps.http.get("https://httpbin.org/get").build();
assert_eq!(req.method(), crux_http::http::Method::Get);

pub fn url(&self) -> &Url

Get the request url.

Examples

use crux_http::http::Url;
let req = caps.http.get("https://httpbin.org/get").build();
assert_eq!(req.url(), &Url::parse("https://httpbin.org/get")?);

pub fn content_type(&self) -> Option<Mime>

Get the request content type as a Mime.

Gets the Content-Type header and parses it to a Mime type.

Read more on MDN

Panics

This method will panic if an invalid MIME type was set as a header. Use the set_header method to bypass any checks.

pub fn set_content_type(&mut self, mime: Mime)

Set the request content type from a Mime.

Read more on MDN

pub fn len(&self) -> Option<usize>

Get the length of the body stream, if it has been set.

This value is set when passing a fixed-size object into as the body. E.g. a string, or a buffer. Consumers of this API should check this value to decide whether to use Chunked encoding, or set the response length.

pub fn is_empty(&self) -> Option<bool>

Returns true if the set length of the body stream is zero, false otherwise.

pub fn set_body(&mut self, body: impl Into<Body>)

Pass an AsyncRead stream as the request body.

Mime

The encoding is set to application/octet-stream.

pub fn take_body(&mut self) -> Body

Take the request body as a Body.

This method can be called after the body has already been taken or read, but will return an empty Body.

This is useful for consuming the body via an AsyncReader or AsyncBufReader.

pub fn body_json(&mut self, json: &impl Serialize) -> Result<()>

Pass JSON as the request body.

Mime

The content-type is set to application/json.

Errors

This method will return an error if the provided data could not be serialized to JSON.

pub fn body_string(&mut self, string: String)

Pass a string as the request body.

Mime

The content-type is set to text/plain; charset=utf-8.

pub fn body_bytes(&mut self, bytes: impl AsRef<[u8]>)

Pass bytes as the request body.

Mime

The content-type is set to application/octet-stream.

pub async fn body_file(&mut self, path: impl AsRef<Path>) -> Result<()>

Pass a file as the request body.

Mime

The content-type is set based on the file extension using mime_guess if the operation was successful. If path has no extension, or its extension has no known MIME type mapping, then None is returned.

Errors

This method will return an error if the file couldn’t be read.

pub fn body_form(&mut self, form: &impl Serialize) -> Result<()>

Pass a form as the request body.

Mime

The content-type is set to application/x-www-form-urlencoded.

Errors

An error will be returned if the encoding failed.

pub fn middleware(&mut self, middleware: impl Middleware)

Push middleware onto a per-request middleware stack.

Important: Setting per-request middleware incurs extra allocations. Creating a Client with middleware is recommended.

Client middleware is run before per-request middleware.

See the middleware submodule for more information on middleware.

Examples

let mut req = caps.http.get("https://httpbin.org/get").build();
req.middleware(crux_http::middleware::Redirect::default());