Struct crux_http::Response

pub struct Response<Body> { /* private fields */ }

An HTTP Response that will be passed to in a message to an apps update function

Implementations

impl<Body> Response<Body>

pub fn status(&self) -> StatusCode

Get the HTTP status code.

Examples

assert_eq!(res.status(), 200);

pub fn version(&self) -> Option<Version>

Get the HTTP protocol version.

Examples

use crux_http::http::Version;
assert_eq!(res.version(), Some(Version::Http1_1));

pub fn header(&self, name: impl Into<HeaderName>) -> Option<&HeaderValues>

Get a header.

Examples

assert!(res.header("Content-Length").is_some());

pub fn header_mut(
&mut self,
name: impl Into<HeaderName>
) -> Option<&mut HeaderValues>

Get an HTTP header mutably.

pub fn remove_header(
&mut self,
name: impl Into<HeaderName>
) -> Option<HeaderValues>

Remove a header.

pub fn insert_header(
&mut self,
key: impl Into<HeaderName>,
value: impl ToHeaderValues
)

Insert an HTTP header.

pub fn append_header(
&mut self,
key: impl Into<HeaderName>,
value: impl ToHeaderValues
)

Append an HTTP header.

pub fn iter(&self) -> Iter<'_>

An iterator visiting all header pairs in arbitrary order.

pub fn iter_mut(&mut self) -> IterMut<'_>

An iterator visiting all header pairs in arbitrary order, with mutable references to the values.

pub fn header_names(&self) -> Names<'_>

An iterator visiting all header names in arbitrary order.

pub fn header_values(&self) -> Values<'_>

An iterator visiting all header values in arbitrary order.

pub fn content_type(&self) -> Option<Mime>

Get the response content type as a Mime.

Gets the Content-Type header and parses it to a Mime type.

Read more on MDN

Panics

This method will panic if an invalid MIME type was set as a header.

Examples

use crux_http::http::mime;
assert_eq!(res.content_type(), Some(mime::JSON));

pub fn body(&self) -> Option<&Body>

pub fn take_body(&mut self) -> Option<Body>

pub fn with_body<NewBody>(self, body: NewBody) -> Response<NewBody>

impl Response<Vec<u8>>

pub fn body_bytes(&mut self) -> Result<Vec<u8>>

Reads the entire request body into a byte buffer.

This method can be called after the body has already been read, but will produce an empty buffer.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

Examples

let bytes: Vec<u8> = res.body_bytes()?;

pub fn body_string(&mut self) -> Result<String>

Reads the entire response body into a string.

This method can be called after the body has already been read, but will produce an empty buffer.

Encodings

If the “encoding” feature is enabled, this method tries to decode the body with the encoding that is specified in the Content-Type header. If the header does not specify an encoding, UTF-8 is assumed. If the “encoding” feature is disabled, Surf only supports reading UTF-8 response bodies. The “encoding” feature is enabled by default.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

If the body cannot be interpreted because the encoding is unsupported or incorrect, an Err is returned.

Examples

let string: String = res.body_string()?;
assert_eq!(string, "hello");

pub fn body_json<T: DeserializeOwned>(&mut self) -> Result<T>

Reads and deserialized the entire request body from json.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

If the body cannot be interpreted as valid json for the target type T, an Err is returned.

Examples

#[derive(Deserialize, Serialize)]
struct Ip {
    ip: String
}

let Ip { ip } = res.body_json()?;
assert_eq!(ip, "127.0.0.1");