crux_http
Struct ResponseAsync

pub struct ResponseAsync { /* private fields */ }

An HTTP response that exposes async methods. This is to support async use and middleware.

Implementations

impl ResponseAsync

pub fn status(&self) -> StatusCode

Get the HTTP status code.

Examples

let res = client.get("https://httpbin.org/get").await?;
assert_eq!(res.status(), 200);

pub fn version(&self) -> Option<Version>

Get the HTTP protocol version.

Examples

use crux_http::http::Version;

let res = client.get("https://httpbin.org/get").await?;
assert_eq!(res.version(), Some(Version::Http1_1));

pub fn header(&self, name: impl Into<HeaderName>) -> Option<&HeaderValues>

Get a header.

Examples

let res = client.get("https://httpbin.org/get").await?;
assert!(res.header("Content-Length").is_some());

pub fn header_mut( &mut self, name: impl Into<HeaderName>, ) -> Option<&mut HeaderValues>

Get an HTTP header mutably.

pub fn remove_header( &mut self, name: impl Into<HeaderName>, ) -> Option<HeaderValues>

Remove a header.

pub fn insert_header( &mut self, key: impl Into<HeaderName>, value: impl ToHeaderValues, )

Insert an HTTP header.

pub fn append_header( &mut self, key: impl Into<HeaderName>, value: impl ToHeaderValues, )

Append an HTTP header.

pub fn iter(&self) -> Iter<'_>

An iterator visiting all header pairs in arbitrary order.

pub fn iter_mut(&mut self) -> IterMut<'_>

An iterator visiting all header pairs in arbitrary order, with mutable references to the values.

pub fn header_names(&self) -> Names<'_>

An iterator visiting all header names in arbitrary order.

pub fn header_values(&self) -> Values<'_>

An iterator visiting all header values in arbitrary order.

pub fn ext<T: Send + Sync + 'static>(&self) -> Option<&T>

Get a response scoped extension value.

pub fn insert_ext<T: Send + Sync + 'static>(&mut self, val: T)

Set a response scoped extension value.

pub fn content_type(&self) -> Option<Mime>

Get the response content type as a Mime.

Gets the Content-Type header and parses it to a Mime type.

Read more on MDN

Panics

This method will panic if an invalid MIME type was set as a header.

Examples

use crux_http::http::mime;
let res = client.get("https://httpbin.org/json").await?;
assert_eq!(res.content_type(), Some(mime::JSON));

pub fn len(&self) -> Option<usize>

Get the length of the body stream, if it has been set.

This value is set when passing a fixed-size object into as the body. E.g. a string, or a buffer. Consumers of this API should check this value to decide whether to use Chunked encoding, or set the response length.

pub fn is_empty(&self) -> Option<bool>

Returns true if the set length of the body stream is zero, false otherwise.

pub fn set_body(&mut self, body: impl Into<Body>)

Set the body reader.

pub fn take_body(&mut self) -> Body

Take the response body as a Body.

This method can be called after the body has already been taken or read, but will return an empty Body.

Useful for adjusting the whole body, such as in middleware.

pub fn swap_body(&mut self, body: &mut Body)

Swaps the value of the body with another body, without deinitializing either one.

pub async fn body_bytes(&mut self) -> Result<Vec<u8>>

Reads the entire request body into a byte buffer.

This method can be called after the body has already been read, but will produce an empty buffer.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

Examples

let mut res = client.get("https://httpbin.org/get").await?;
let bytes: Vec<u8> = res.body_bytes().await?;

pub async fn body_string(&mut self) -> Result<String>

Reads the entire response body into a string.

This method can be called after the body has already been read, but will produce an empty buffer.

Encodings

If the “encoding” feature is enabled, this method tries to decode the body with the encoding that is specified in the Content-Type header. If the header does not specify an encoding, UTF-8 is assumed. If the “encoding” feature is disabled, Surf only supports reading UTF-8 response bodies. The “encoding” feature is enabled by default.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

If the body cannot be interpreted because the encoding is unsupported or incorrect, an Err is returned.

Examples

let mut res = client.get("https://httpbin.org/get").await?;
let string: String = res.body_string().await?;

pub async fn body_json<T: DeserializeOwned>(&mut self) -> Result<T>

Reads and deserialized the entire request body from json.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

If the body cannot be interpreted as valid json for the target type T, an Err is returned.

Examples

#[derive(Deserialize, Serialize)]
struct Ip {
    ip: String
}

let mut res = client.get("https://api.ipify.org?format=json").await?;
let Ip { ip } = res.body_json().await?;

pub async fn body_form<T: DeserializeOwned>(&mut self) -> Result<T>

Reads and deserialized the entire request body from form encoding.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

If the body cannot be interpreted as valid json for the target type T, an Err is returned.

Examples

#[derive(Deserialize, Serialize)]
struct Body {
    apples: u32
}

let mut res = client.get("https://api.example.com/v1/response").await?;
let Body { apples } = res.body_form().await?;