Crate crux_http

A HTTP client for use with Crux

crux_http allows Crux apps to make HTTP requests by asking the Shell to perform them.

This is still work in progress and large parts of HTTP are not yet supported.

Re-exports

pub use http_types as http;

Modules

client

middleware

Middleware types

protocol

The protocol for communicating with the shell

testing

Structs

Config

Configuration for crux_http::Https and their underlying HTTP client.

Http

The Http capability API.

Request

An HTTP request, returns a Response.

RequestBuilder

Request Builder

Response

An HTTP Response that will be passed to in a message to an apps update function

ResponseAsync

An HTTP response that exposes async methods. This is to support async use and middleware.

Enums

HttpError

Type Aliases

Result