crux_http::client
Struct Client

pub struct Client { /* private fields */ }

An HTTP client, capable of sending Requests

Users should only interact with this type from middlewares - normal crux code should make use of the Http capability type instead.

Examples

use futures_util::future::BoxFuture;
use crux_http::middleware::{Next, Middleware};
use crux_http::{client::Client, Request, RequestBuilder, ResponseAsync, Result};
use std::time;
use std::sync::Arc;

// Fetches an authorization token prior to making a request
fn fetch_auth<'a>(mut req: Request, client: Client, next: Next<'a>) -> BoxFuture<'a, Result<ResponseAsync>> {
    Box::pin(async move {
        let auth_token = client.get("https://httpbin.org/get")
            .await?
            .body_string()
            .await?;
        req.append_header("Authorization", format!("Bearer {auth_token}"));
        next.run(req, client).await
    })
}

Implementations

impl Client

pub async fn send(&self, req: impl Into<Request>) -> Result<ResponseAsync>

Send a Request using this client.

pub async fn recv_bytes(&self, req: impl Into<Request>) -> Result<Vec<u8>>

Submit a Request and get the response body as bytes.

pub async fn recv_string(&self, req: impl Into<Request>) -> Result<String>

Submit a Request and get the response body as a string.

pub async fn recv_json<T: DeserializeOwned>( &self, req: impl Into<Request>, ) -> Result<T>

Submit a Request and decode the response body from json into a struct.

pub async fn recv_form<T: DeserializeOwned>( &self, req: impl Into<Request>, ) -> Result<T>

Submit a Request and decode the response body from form encoding into a struct.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

If the body cannot be interpreted as valid json for the target type T, an Err is returned.

pub fn get(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP GET request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn head(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP HEAD request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn post(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP POST request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn put(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP PUT request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn delete(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP DELETE request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn connect(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP CONNECT request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn options(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP OPTIONS request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn trace(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP TRACE request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn patch(&self, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform an HTTP PATCH request using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn request(&self, verb: Method, uri: impl AsRef<str>) -> RequestBuilder<()>

Perform a HTTP request with the given verb using the Client connection.

Panics

This will panic if a malformed URL is passed.

Errors

Returns errors from the middleware, http backend, and network sockets.

pub fn config(&self) -> &Config

Get the current configuration.